Folded in some partials that were only used in one place.
